+++ Backend.Application/DTOs/StockTakes/StockTakeItemDto.cs
@@ -8,9 +8,9 @@ public class StockTakeItemDto
     public int StockTakeId { get; set; }
     public int? ProductVariantId { get; set; }
     public int? LocationId { get; set; }
-    public int SystemQuantity { get; set; }
-    public int? ActualQuantity { get; set; }
-    public int Difference { get; set; }
+    public decimal SystemQuantity { get; set; }
+    public decimal? ActualQuantity { get; set; }
+    public decimal Difference { get; set; }
     public string? Note { get; set; }
     public bool QRScanned { get; set; }
 }
@@ -19,8 +19,8 @@ public class CreateStockTakeItemDto
 {
     public int? ProductVariantId { get; set; }
     public int? LocationId { get; set; }
-    public int SystemQuantity { get; set; }
-    public int? ActualQuantity { get; set; }
+    public decimal SystemQuantity { get; set; }
+    public decimal? ActualQuantity { get; set; }
     public string? Note { get; set; }
     public bool QRScanned { get; set; }
 }
@@ -29,3 +29,4 @@ public class UpdateStockTakeItemDto : CreateStockTakeItemDto
 {
     public int Id { get; set; }
 }
+
